Pull request #62: MEDDV002-6444; Refinement of data restrictions
Merge in MED/endava-marketplace from feature/dm-refinement to development

Squashed commit of the following:

commit f33f3725149d6d24cf569be64ef99bee0ec07bda
Author: fmontoya <federico.montoya@endava.com>
Date:   Thu Oct 26 11:33:33 2023 -0500

    Refinement of data restrictions

+++ backend/src/main/java/com/endava/marketplace/backend/controller/EndavanController.java
@@ -43,7 +43,7 @@ public class EndavanController {
             tags = {"Endavan"}
     )
     @GetMapping("/get/{id}")
-    public Optional<Endavan> getEndavanById(@PathVariable Integer id){
+    public Optional<Endavan> getEndavanById(@PathVariable Long id){
         return endavanService.findEndavanById(id);
     }
     @Operation(
@@ -52,7 +52,7 @@ public class EndavanController {
             tags = {"Endavan"}
     )
     @DeleteMapping("/delete/{id}")
-    public void deleteEndavanById(@PathVariable Integer id){
+    public void deleteEndavanById(@PathVariable Long id){
         endavanService.deleteEndavanById(id);
     }
