+++ backend/src/main/java/com/endava/marketplace/backend/repository/ListingCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ListingCategoryRepository extends JpaRepository<ListingCategory, Integer> {
+public interface ListingCategoryRepository extends JpaRepository<ListingCategory, Long> {
 }
